Permissions

+++ Modules/Backend/Resources/views/users/index.blade.php
@@ -29,6 +29,7 @@
                             <th>S.No</th>
                             <th>Full Name</th>
                             <th>Email</th>
+                            <th>Role</th>
                             <th>Image</th>
                             <th>Status</th>
                             <th class="no-sort">Action</th>
@@ -36,25 +37,36 @@
                         </thead>
                         <tbody>
                         @foreach($users as $user)
+
                             <tr>
                                 <td>{{$user->id}}</td>
-                                <td>{{$user->full_name}}</td>
+                                <td>{{$user->user_name}}</td>
                                 <td>{{$user->email}}</td>
+                                <td>
+                                    @if(count($user->roles))
+                                        <span class="label label-primary">
+                                        {{$user->getRoleNames()->first()}}
+                                    </span>
+                                    @endif
+                                </td>
                                 <td>{{$user->image}}</td>
                                 <td>{!! spanByStatus($user->status) !!}</td>
                                 <td>
-                                    {{--                                    <a href="{{route($routePrefix.'.users.edit',$user->id)}}"--}}
-                                    {{--                                       class="btn btn-primary btn-sm btn-flat">--}}
-                                    {{--                                        <i class="fa fa-edit"></i>--}}
-                                    {{--                                    </a>--}}
-                                    {!! Form::open(['method' => 'DELETE', 'route' =>
-                                            [$routePrefix.'users.destroy',$user->id],
-                                             'onsubmit' => "return confirm('Are you sure you want to delete?')",   'style'=>"display:inline"
-                                       ])!!}
-                                    <button class="btn btn-danger btn-flat btn-sm" role="button" type="submit">
-                                        <i class="fa fa-trash"></i>
-                                    </button>
-                                    {!! Form::close() !!}
+
+                                    @if(!$user->isSuper())
+                                        <a href="{{route($routePrefix.'users.edit',$user->id)}}"
+                                           class="btn btn-primary btn-sm btn-flat">
+                                            <i class="fa fa-edit"></i>
+                                        </a>
+                                        {!! Form::open(['method' => 'DELETE', 'route' =>
+                                                [$routePrefix.'users.destroy',$user->id],
+                                                 'onsubmit' => "return confirm('Are you sure you want to delete?')",   'style'=>"display:inline"
+                                           ])!!}
+                                        <button class="btn btn-danger btn-flat btn-sm" role="button" type="submit">
+                                            <i class="fa fa-trash"></i>
+                                        </button>
+                                        {!! Form::close() !!}
+                                    @endif
                                 </td>
                             </tr>
                         @endforeach
